requirements.win.txt
glob2==0.7
joblib==1.4.2
lightgbm==4.3.0
matplotlib==3.8.4
matplotlib-inline==0.1.6
numpy==1.26.4
opencv-python-headless==4.9.0.80
pandas==2.2.2
pathlib==1.0.1
pillow==10.4.0
protobuf==3.20.3
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.11.4
tqdm==4.66.5
xgboost==2.1.1
h5py==3.11.0

# TensorFlow + DirectML (Windows GPU backend)
tensorflow-directml-plugin


